Refactor auth and upload controllers
